[TARGETING] Add stopwatch timing for visitorInfo, condition matching and data loading
